fix: dont show tracking info unless requested
