+++ lib/init.js
@@ -38,7 +38,7 @@ function _init (opts, tracker, cb) {
       return cbPassedIn ? cb(err) : setImmediate(() => tracker.emit('error', err))
     }
 
-    setImmediate(() => tracker.emit('done', result))
+    tracker.emit('done', result)
     cb(null, result)
 
     if (!err && opts.json) {
