+++ lib/manager.js
@@ -4,7 +4,6 @@ const path = require('path')
 const EE = require('events').EventEmitter
 const aggregateResult = require('./aggregateResult')
 const { getHistograms } = require('./histUtil')
-const ora = require('ora')
 const { Worker } = require('./worker_threads')
 
 function initWorkers (opts, tracker, cb) {
@@ -19,7 +18,6 @@ function initWorkers (opts, tracker, cb) {
     throughput: []
   }
   let restart = true
-  let spinner
 
   function reset () {
     workers.length = 0
@@ -30,29 +28,15 @@ function initWorkers (opts, tracker, cb) {
     histData.throughput.length = 0
   }
 
-  function showSpinner () {
-    spinner = ora('running...')
-    spinner.start()
-  }
-
-  function hideSpinner () {
-    if (spinner) {
-      spinner.stop()
-      spinner = null
-    }
-  }
-
   function startAll () {
-    logToStream(`Running ${opts.duration}s test @ ${opts.url}\n${opts.connections} connections\n${numWorkers} workers\n`)
-    if (process.stderr.isTTY) showSpinner()
-
     for (const w of workers) {
       w.postMessage({ cmd: 'START' })
     }
+
+    setImmediate(() => { tracker.emit('start') })
   }
 
   function handleFinish () {
-    hideSpinner()
     const result = aggregateResult(results, opts, histograms)
     cb(null, result)
     reset()
@@ -66,16 +50,10 @@ function initWorkers (opts, tracker, cb) {
   }
 
   function stopAll () {
-    hideSpinner()
-
     for (const w of workers) {
       w.postMessage({ cmd: 'STOP' })
     }
     reset()
-  }
-
-  function logToStream (msg) {
-    process.stderr.write(msg + '\n')
   }
 
   const workerOpts = {
