+++ lib/progressTracker.js
@@ -8,6 +8,7 @@
 const ProgressBar = require('progress')
 const Chalk = require('chalk')
 const testColorSupport = require('color-support')
+const ora = require('ora')
 const printResult = require('./printResult')
 const { isMainThread } = require('./worker_threads')
 const defaults = {
@@ -33,6 +34,7 @@ function track (instance, opts) {
   let durationProgressBar
   let amountProgressBar
   let addedListeners = false
+  let spinner
 
   instance.on('start', () => {
     if (opts.renderProgressBar && isMainThread) {
@@ -43,14 +45,24 @@ function track (instance, opts) {
         msg += ` with ${iOpts.pipelining} pipelining factor`
       }
 
-      if (!iOpts.amount) {
-        logToStream(`Running ${iOpts.duration}s test @ ${iOpts.url}${socketPath}\n${msg}\n`)
+      if (iOpts.workers) {
+        msg += `\n${iOpts.workers} workers`
+      }
 
-        durationProgressBar = trackDuration(instance, opts, iOpts)
+      const message = iOpts.amount
+        ? `Running ${iOpts.amount} requests test @ ${iOpts.url}${socketPath}\n${msg}\n`
+        : `Running ${iOpts.duration}s test @ ${iOpts.url}${socketPath}\n${msg}\n`
+
+      logToStream(message)
+
+      if (iOpts.workers) {
+        showSpinner()
       } else {
-        logToStream(`Running ${iOpts.amount} requests test @ ${iOpts.url}${socketPath}\n${msg}\n`)
-
-        amountProgressBar = trackAmount(instance, opts, iOpts)
+        if (iOpts.amount) {
+          amountProgressBar = trackAmount(instance, opts, iOpts)
+        } else {
+          durationProgressBar = trackDuration(instance, opts, iOpts)
+        }
       }
 
       addListener()
@@ -66,6 +78,18 @@ function track (instance, opts) {
     opts.outputStream.write(tableStrResult)
   })
 
+  function showSpinner () {
+    spinner = ora('running...')
+    spinner.start()
+  }
+
+  function hideSpinner () {
+    if (spinner) {
+      spinner.stop()
+      spinner = null
+    }
+  }
+
   function addListener () {
     // add listeners for progress bar to instance here so they aren't
     // added on restarting, causing listener leaks
@@ -74,6 +98,18 @@ function track (instance, opts) {
     }
 
     addedListeners = true
+
+    // TODO: think about if workers can report progress every second, that
+    // way we could have a progress bar per worker.
+    if (iOpts.workers) {
+      // using `prependOnceListener` to make sure that we hide the spinner
+      // before writing anything else.
+      // if we print anything else to the output, the stale spinner text
+      // is left uncleared and clutters the output.
+      instance.prependOnceListener('done', hideSpinner)
+      return
+    }
+
     // note: Attempted to curry the functions below, but that breaks the functionality
     // as they use the scope/closure of the progress bar variables to allow them to be reset
     if (opts.outputStream.isTTY) {
